requirements.txt
anthropic>=0.40.0
pymupdf>=1.24.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
click>=8.0.0
sqlalchemy>=2.0.0
tiktoken>=0.7.0
tenacity>=8.0.0
pytest>=7.0.0
fountain>=0.1.0
jinja2>=3.0.0
